+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) starting from 
 - Improved link resolving via the new `Link` type
 - Swift 3.1, Xcode 8.3 support.
 
+#### Changed
+- **BREAKING:** `Contentful.Array` is now called `ArrayResponse` to avoid clashing with native Swift arrays.
 #### Planned
 - Better support and Swift'ier API for [Images API](https://www.contentful.com/developers/docs/references/images-api/).